requirements.txt
certifi==2021.10.8
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pymongo==3.12.1
python-dotenv==0.19.2
Werkzeug==2.0.2
Flask-PyMongo
